Update: make a base class for NCAA units
